Rename Mime namespace to Stream and StreamCopy to HandleToResource
The src/Mime classes are stream/handle bridges, not MIME logic, so the
namespace is renamed to Stream. StreamCopy is renamed to HandleToResource
(it drains a Psl\IO read handle into a resource) with clearer method names:
stream() returns a ResourceStream, resource() returns the raw resource.
